refactor(ai): organize chat module — types/schemas/utils/constants in their folders
Per project convention, types live in types/, schemas in schemas/, helpers in utils/,
and constants in constants/. Move chat sub-files out of the flat chat/ folder into
their proper homes; chat/ now holds only the runtime backend logic (history,
abort-polling, config).
