+++ apps/dashboard/src/components/chat/chat-input-context-row.tsx
@@ -1,6 +1,6 @@
 import { Cancel01Icon, TextSelectionIcon } from "@hugeicons/core-free-icons";
 import { HugeiconsIcon } from "@hugeicons/react";
-import type { ContextItem, TextSelection } from "@notra/ai/chat/types";
+import type { ContextItem, TextSelection } from "@notra/ai/types/chat";
 import { Github } from "@notra/ui/components/ui/svgs/github";
 import { Linear } from "@notra/ui/components/ui/svgs/linear";
 import {
