+++ apps/dashboard/src/components/dashboard/chat-history-nav.tsx
@@ -9,9 +9,9 @@ import {
   PinOffIcon,
 } from "@hugeicons/core-free-icons";
 import { HugeiconsIcon } from "@hugeicons/react";
-import { CHAT_TITLE_MAX_LENGTH } from "@notra/ai/chat/constants";
-import { normalizeChatTitle } from "@notra/ai/chat/metadata";
-import type { ChatSessionSummary } from "@notra/ai/chat/types";
+import { CHAT_TITLE_MAX_LENGTH } from "@notra/ai/constants/chat";
+import { normalizeChatTitle } from "@notra/ai/utils/chat";
+import type { ChatSessionSummary } from "@notra/ai/types/chat";
 import {
   ResponsiveAlertDialog,
   ResponsiveAlertDialogAction,
