+++ apps/dashboard/src/components/dashboard/chat-topbar-title.tsx
@@ -8,11 +8,11 @@ import {
   PinOffIcon,
 } from "@hugeicons/core-free-icons";
 import { HugeiconsIcon } from "@hugeicons/react";
-import { CHAT_TITLE_MAX_LENGTH } from "@notra/ai/chat/constants";
+import { CHAT_TITLE_MAX_LENGTH } from "@notra/ai/constants/chat";
 import {
   formatChatIdFallback,
   normalizeChatTitle,
-} from "@notra/ai/chat/metadata";
+} from "@notra/ai/utils/chat";
 import {
   ResponsiveAlertDialog,
   ResponsiveAlertDialogAction,
