+++ apps/dashboard/src/lib/hooks/use-chat-sessions.ts
@@ -5,12 +5,12 @@ import {
   chatSessionsPath,
   chatSessionsQueryKey,
   sortChatSessions,
-} from "@notra/ai/chat/metadata";
+} from "@notra/ai/utils/chat";
 import {
   chatSessionResponseSchema,
   chatSessionsListResponseSchema,
-} from "@notra/ai/chat/schemas";
-import type { ChatSessionSummary } from "@notra/ai/chat/types";
+} from "@notra/ai/schemas/chat";
+import type { ChatSessionSummary } from "@notra/ai/types/chat";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useRef } from "react";
 import { toast } from "sonner";
