+++ packages/ai/src/constants/chat.ts
